Change Source to tarball obtained with obs_scm
The tarball is obtained with the obs_scm service for convenience (easier
to fetch contents of git tags as well as test branches). This Source
line would overwrite the obs_scm tarball with a release tarball
downloaded with the implicit download_files service, thus losing the
ability to handle test branches.